requirements.txt
azureml-sdk
pyyaml